remove deprecae buttonColor and change stateful to stateless

+++ lib/models/list_restaurant_model.dart
@@ -45,7 +45,7 @@ class Restaurants {
         description: json['description'],
         pictureId: json['pictureId'],
         city: json['city'],
-        rating: json['rating']);
+        rating: json['rating'].toDouble());
   }
 
   Map<String, dynamic> toJson() {
